#34: Add profile settings page

+++ mobile/lib/screens/goal/goal_view_screen.dart
@@ -339,7 +339,7 @@ class _GoalViewScreenState extends BaseEditorScreen<GoalViewScreen, Goal> {
 
   void _showDescriptionModal(Size size, Goal goal) => simpleBottomSheet(
         title: 'Add description',
-        height: (size.height * 0.9).toInt(),
+        height: size.height * 0.9,
         child: TextEditModal(
           placeholder: 'Description',
           initialValue: goal.description,
